Implemented ldloc; defaulted on sizeof(void*) and Int32 as (long) for correct sign.

+++ ~/C699.cs
@@ -161,7 +161,8 @@ public struct c
 		}
 	public c Int
 		{
-		get { s += KeyedWord.Int + ' '  ; return this ; }
+		get { s += KeyedWord.Long + ' '  ; return this ; }
+		//get { s += KeyedWord.Int + ' '  ; return this ; }
 		}
 	public c Static(C699.c c, C699.c symbol)
 		{
